sep(heredc): sep create_heredoc_filename() in new file
