+++ Makefile
@@ -77,6 +77,7 @@ SRCS		+=	$(PARSE_DIR)/abstruct_syntax_tree.c \
 				$(PARSE_DIR)/ast_convert_kind.c \
 				$(PARSE_DIR)/ast_dup_token.c \
 				$(PARSE_DIR)/ast_print_error.c \
+				$(PARSE_DIR)/create_heredoc_filename.c \
 				$(PARSE_DIR)/destroy_ast.c \
 				$(PARSE_DIR)/heredoc.c \
 				$(PARSE_DIR)/is_parenthesis_concatted.c \
