+++ srcs/parse/heredoc.c
@@ -12,95 +12,6 @@
 #include "ft_sys.h"
 #include "ft_mem.h"
 
-static int	count_up_seed_num(void)
-{
-	static int	num; //todo: init value?
-
-	num++;
-	return (num);
-}
-
-static char *x_ft_itoa(int n)
-{
-	char	*num_str;
-
-	num_str = ft_itoa(n);
-	if (!num_str)
-		ft_abort();
-	return (num_str);
-}
-
-static char *x_ft_strjoin(char const *s1, char const *s2)
-{
-	char	*join;
-
-	join = ft_strjoin(s1, s2);
-	if (!join)
-		ft_abort();
-	return (join);
-}
-
-static char	*create_hidden_filename(const char *hash_str, const char *seed_str)
-{
-	char	*filename;
-	char	*hidden_filename;
-
-	filename = x_ft_strjoin(hash_str, seed_str);
-	hidden_filename = x_ft_strjoin(HEREDOC_FILE_PREFIX, filename);
-	ft_free(&filename);
-	return (hidden_filename);
-}
-
-static char	*create_hidden_filename_based_on_hash(void)
-{
-	const int	seed = count_up_seed_num();
-	size_t		hash;
-	char 		*seed_str;
-	char 		*hash_str;
-	char 		*hidden_filename;
-
-	seed_str = x_ft_itoa(seed);
-	hash = hs_gen_fnv((const unsigned char*)seed_str, (size_t)1);
-	hash_str = x_ft_itoa((int)hash);
-	hidden_filename = create_hidden_filename(hash_str, seed_str);
-	ft_free(&seed_str);
-	ft_free(&hash_str);
-	return (hidden_filename);
-}
-
-// file can create when result of access(2) is no such file
-static bool	is_file_creatable(const char *filepath)
-{
-	int	tmp_err;
-	int	ret;
-
-	errno = 0;
-	ret = access(filepath, F_OK);
-	tmp_err = errno;
-	if (ret == ACCESS_ERROR && tmp_err == ENOENT)
-		return (true);
-	return (false);
-}
-
-// filename = <hash> + num -> itoa
-static char *create_heredoc_filename(void)
-{
-	char	*filename;
-
-	while (true)
-	{
-		filename = create_hidden_filename_based_on_hash();
-//		ft_dprintf(2, "try access filename:%s\n", filename);
-		if (is_file_creatable(filename))
-		{
-//			ft_dprintf(2, " -> filename:%s\n", filename);
-			break ;
-		}
-		ft_free(&filename);
-	}
-	return (filename);
-}
-
 static void	clean_up_prev_heredoc(int in_fd, char *filename)
 {
 	close(in_fd); 		// todo: error handling
